parameterize populate.py script and fix pydicom import
Instead of using hard-coded paths locations, reuse those from settings. While doing that,
added XML annotations location to the configuration. It might make sense to move the getSettings()
function to the util class. It might also be good to add the path to the sqlite database
to the configuration, using the installed one by default.

Fixed `import dicom`, which was removed in pydicom 1.0 - now this is handeled more gracefully
by catching ImportError.